@@ -1,10 +1,31 @@
+import java.io.FileInputStream
+import java.util.Properties
+
 plugins {
     alias(libs.plugins.android.application)
+    alias(libs.plugins.kotlin.android)
+    alias(libs.plugins.kotlin.compose)
+}
+
+fun getIpAddress(): String? {
+    val properties = Properties()
+    val localPropertiesFile = File("local.properties")
+
+    try {
+        FileInputStream(localPropertiesFile).use { inputStream ->
+            properties.load(inputStream)
+        }
+        return properties.getProperty("ip_addr") // Replace "ipAddress" with the key you want to retrieve
+    } catch (e: Exception) {
+        e.printStackTrace()
+    }
+
+    return null
 }
 
 android {
     namespace = "com.example.eventure"
-    compileSdk = 34
+    compileSdk = 35
 
     defaultConfig {
         applicationId = "com.example.eventure"
@@ -12,7 +33,9 @@ android {
         targetSdk = 34
         versionCode = 1
         versionName = "1.0"
-
+        val ipAddress = getIpAddress()
+        buildConfigField("String", "IP_ADDR", "\"$ipAddress\"")
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
     }
 
@@ -31,6 +54,8 @@ android {
     }
     buildFeatures {
         viewBinding = true
+        buildConfig = true
+        compose = true
     }
 
 }
@@ -43,9 +68,39 @@ dependencies {
     implementation(libs.constraintlayout)
     implementation(libs.navigation.fragment)
     implementation(libs.navigation.ui)
+    implementation(libs.lifecycle.runtime.ktx)
+    implementation(libs.activity.compose)
+    implementation(platform(libs.compose.bom))
+    implementation(libs.ui)
+    implementation(libs.ui.graphics)
+    implementation(libs.ui.tooling.preview)
+    implementation(libs.material3)
     testImplementation(libs.junit)
     androidTestImplementation(libs.ext.junit)
     androidTestImplementation(libs.espresso.core)
     implementation(libs.viewpager2)
     implementation(libs.recyclerview)
+    implementation("com.squareup.retrofit2:retrofit:2.9.0")
+    implementation("com.squareup.retrofit2:converter-gson:2.9.0")
+    implementation("com.squareup.okhttp3:logging-interceptor:4.12.0")
+    implementation("com.google.code.gson:gson:2.8.8")
+    implementation("com.github.bumptech.glide:glide:4.15.1")
+    implementation("androidx.paging:paging-runtime:2.1.2")
+    implementation("androidx.lifecycle:lifecycle-extensions:2.2.0")
+    androidTestImplementation(platform(libs.compose.bom))
+    androidTestImplementation(libs.ui.test.junit4)
+    annotationProcessor("com.github.bumptech.glide:compiler:4.15.1")
+    implementation("com.auth0.android:jwtdecode:2.0.0")
+    implementation("com.squareup.okhttp3:okhttp:4.12.0")
+    implementation("com.github.NaikSoftware:StompProtocolAndroid:1.6.6")
+    implementation("io.reactivex.rxjava2:rxjava:2.2.21")
+    implementation("io.reactivex.rxjava2:rxandroid:2.1.1")
+    implementation("com.google.code.gson:gson:2.10.1")
+    implementation ("com.squareup.retrofit2:converter-scalars:2.9.0")
+    implementation ("com.github.bumptech.glide:glide:4.15.1")
+    annotationProcessor ("com.github.bumptech.glide:compiler:4.15.1")
+    debugImplementation(libs.ui.tooling)
+    debugImplementation(libs.ui.test.manifest)
+    implementation ("com.github.PhilJay:MPAndroidChart:v3.1.0")
+    implementation ("com.kizitonwose.calendar:view:2.3.0")
 }
