@@ -10,31 +10,52 @@ import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.eventure.R;
+import com.example.eventure.clients.ClientUtils;
 import com.example.eventure.fragments.LoginFragment;
 
 import java.util.Timer;
 import java.util.TimerTask;
 
 public class SplashScreenActivity extends AppCompatActivity {
-    private static final int SPLASH_SCREEN_DELAY = 5000; //5 second delay
+    private static final int SPLASH_SCREEN_DELAY = 1500; //1,5 second delay
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
+        ClientUtils.initializeAuthService(this);
+
         setContentView(R.layout.activity_splash);
         TextView textView = findViewById(R.id.textView);
         Animation fadeIn = AnimationUtils.loadAnimation(this, R.anim.fade_in);
-        fadeIn.setDuration(4000);
+        fadeIn.setDuration(1000);
         textView.startAnimation(fadeIn);
+
         new Timer().schedule(new TimerTask() {
             @Override
             public void run() {
-                Intent intent = new Intent(SplashScreenActivity.this, ProfileActivity.class);
+                Intent intent = new Intent(SplashScreenActivity.this, HomeActivity.class);
+
+                // Forward deep link data (for quick registration)
+                Intent incomingIntent = getIntent();
+                if (incomingIntent != null && incomingIntent.getData() != null) {
+                    intent.setData(incomingIntent.getData());
+                }
+
                 startActivity(intent);
                 finish();
             }
         }, SPLASH_SCREEN_DELAY);
-
     }
+        /*
+        new Timer().schedule(new TimerTask() {
+            @Override
+            public void run() {
+
+                Intent intent = new Intent(SplashScreenActivity.this,  HomeActivity.class);
+                startActivity(intent);
+                finish();
+            }
+        }, SPLASH_SCREEN_DELAY);
+*/
 }
