@@ -1,66 +1,116 @@
 package com.example.eventure.fragments;
 
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.eventure.R;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link FavoriteProductsFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
+import com.example.eventure.R;
+import com.example.eventure.adapters.OfferAdapter;
+import com.example.eventure.clients.ClientUtils;
+import com.example.eventure.dto.OfferDTO;
+import com.example.eventure.model.PagedResponse;
+
+import java.util.List;
+
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
+
 public class FavoriteProductsFragment extends Fragment {
 
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
+    private RecyclerView offerRecyclerView;
+    private OfferAdapter offerAdapter;
+    private View loadMoreButton;
+    private View emptyOffers;
 
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
+    private int currentPage = 0;
+    private int totalItemsCount = 1;
+    private boolean isLoading = false;
 
     public FavoriteProductsFragment() {
-        // Required empty public constructor
+        // Required empty constructor
     }
 
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment FavoriteProductsFragment.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static FavoriteProductsFragment newInstance(String param1, String param2) {
-        FavoriteProductsFragment fragment = new FavoriteProductsFragment();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
+    public static FavoriteProductsFragment newInstance() {
+        return new FavoriteProductsFragment();
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_favorite_products, container, false);
+        View rootView = inflater.inflate(R.layout.fragment_favorite_offers, container, false);
+
+        offerRecyclerView = rootView.findViewById(R.id.productRecyclerView);
+        emptyOffers = rootView.findViewById(R.id.emptyOffers);
+        loadMoreButton = rootView.findViewById(R.id.loadMoreOffers);
+
+        offerAdapter = new OfferAdapter(getChildFragmentManager());
+        offerRecyclerView.setAdapter(offerAdapter);
+        offerRecyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
+        offerRecyclerView.setNestedScrollingEnabled(false);
+
+        fetchFavoriteProducts(currentPage);
+
+        loadMoreButton.setOnClickListener(v -> {
+            if (!isLoading && offerAdapter.getItemCount() < totalItemsCount) {
+                currentPage++;
+                fetchFavoriteProducts(currentPage);
+            }
+        });
+
+        return rootView;
+    }
+
+    private void fetchFavoriteProducts(int page) {
+        isLoading = true;
+        ClientUtils.userService.getFavoriteProducts(page, ClientUtils.PAGE_SIZE).enqueue(new Callback<PagedResponse<OfferDTO>>() {
+            @Override
+            public void onResponse(Call<PagedResponse<OfferDTO>> call, Response<PagedResponse<OfferDTO>> response) {
+                isLoading = false;
+                if (response.isSuccessful() && response.body() != null) {
+                    List<OfferDTO> offers = response.body().getContent();
+                    totalItemsCount = response.body().getTotalElements();
+
+                    offerAdapter.addOffers(offers);
+                    offerAdapter.notifyItemRangeInserted(
+                            offerAdapter.getItemCount() - offers.size(),
+                            offers.size()
+                    );
+
+                    updateLoadMoreVisibility();
+                    toggleEmptyOffers();
+                } else {
+                    Log.e("FavoriteProducts", "Failed response: " + response.code());
+                }
+            }
+
+            @Override
+            public void onFailure(Call<PagedResponse<OfferDTO>> call, Throwable t) {
+                isLoading = false;
+                Log.e("FavoriteProducts", "Error: ", t);
+            }
+        });
+    }
+
+    private void updateLoadMoreVisibility() {
+        if (offerAdapter.getItemCount() < totalItemsCount) {
+            loadMoreButton.setVisibility(View.VISIBLE);
+        } else {
+            loadMoreButton.setVisibility(View.GONE);
+        }
+    }
+
+    private void toggleEmptyOffers() {
+        if (offerAdapter.getItemCount() == 0) {
+            emptyOffers.setVisibility(View.VISIBLE);
+        } else {
+            emptyOffers.setVisibility(View.GONE);
+        }
     }
 }
