@@ -32,7 +32,7 @@ public class ProfileStartFragment extends Fragment {
             // Find the NavController from the current view
             NavController navController = Navigation.findNavController(v);
             // Navigate to the LoginFragment using the defined action in the navigation graph
-            navController.navigate(R.id.action_profileStartFragment_to_loginFragment);
+            navController.navigate(R.id.loginFragment);
         });
 
         // When the Register button is clicked, navigate to RegisterFragment
@@ -40,7 +40,7 @@ public class ProfileStartFragment extends Fragment {
             // Find the NavController from the current view
             NavController navController = Navigation.findNavController(v);
             // Navigate to the RegisterFragment using the defined action in the navigation graph
-            navController.navigate(R.id.action_profileStartFragment_to_registerFragment);
+            navController.navigate(R.id.registerFragment);
         });
 
         return view;
